Support transferList
When using the module using threads, it's now possible to specify a transferList as thrid argument to the callback. As a consequence there's a breaking change so that when using the child_process version, the third argument to the callback is ignored and no longer sent to the parent process.